requirements.txt
telethon>=1.34.0
pyrogram>=2.0.106
python-dotenv>=1.0.0
aiofiles>=23.2.1
aiohttp>=3.9.1
youtube-dl>=2021.12.17
yt-dlp>=2023.12.30
pydub>=0.25.1
requests>=2.31.0
pillow>=10.1.0
psutil>=5.9.6
gitpython>=3.1.40
colorlog>=6.8.0
asyncio-mqtt>=0.16.1
python-magic>=0.4.27